src/routes/signup.ts
import express, { Request, Response } from "express";
import {
  BadRequestError,
  Token,
  validateRequest,
} from "@4kprojects/project-util-server";
import { User } from "../models";
import { body } from "express-validator";

const router = express.Router();

router.post(
  "/api/users/signup",
  [
    body("email").isEmail().withMessage("Email must be valid"),
    body("password")
      .trim()
      .isLength({ min: 4, max: 20 })
      .withMessage("Password must be between 4 and 20 characters"),
  ],
  validateRequest,
  async (req: Request, res: Response) => {
    const { email, username = email, password } = req.body;

    const existingUser = await User.findOne({ email });

    if (existingUser) {
      throw new BadRequestError("Email in use");
    }

    const user = User.build({ email, username, password });
    await user.save();

    const userJwt = Token.tokenizeUser(user as UserPayload);

    req.session = {
      jwt: userJwt,
    };

    res.status(201).send(user);
  }
);

export { router as signupRouter };
